Moved directory scanning logic entirely out to dirwatch; makes more sense to split this off so the core file is not like 2k lines long (and it's all related).
